Update:小修改

+++ GameServer/AscensionProtocol/DTO/CricketStatusDTO.cs
@@ -22,6 +22,6 @@ namespace AscensionProtocol
         public virtual int ReduceDef { set; get; }
         public virtual int Rebound { set; get; }
         public virtual int Eva { set; get; }
-        public virtual int Speed { set; get; }
+        public virtual double Speed { set; get; }
     }
 }
